+++ configuration.h
@@ -110,6 +110,10 @@ typedef struct settings
       /* Audio */
       bool audio_enable;
       bool audio_enable_menu;
+      bool audio_enable_menu_ok;
+      bool audio_enable_menu_cancel;
+      bool audio_enable_menu_notice;
+      bool audio_enable_menu_bgm;
       bool audio_sync;
       bool audio_rate_control;
       bool audio_wasapi_exclusive_mode;